+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Enforce JWT on API side
+- Only return results by user ID
 
 ## [0.1.2] - 2019-04-15
 ### Added